Cleaned up srcDiffRunner's comments and removed the functions that allowed for srcDiff to receive direct xml input on the command line through stdin. Also added in the copyright.
Finally, I added in a srcToolsHelper class to provide the same functionality as the SrcMLHelper does except this one is for tools that run on top of srcML (srcSlice, srcType, etc). It can be used to generate paths to the tools based on which one you want at any given point. Currently untested, however.