Updated tests to work with newer version of srcML. This mostly involved updating assumptions about srcML's grammar to fit how its been modified. It also involved using LINQ to deal with checking for if all units were translated from source to srcML.
A note on that latter fix: There seemed to be an assumption about the order in which units were placed in the srcML archive. I assumed this order didn't really matter much (as there are only a few reasons why you'd need them in a particular order). In the spirit of this, I used LINQ to just check and see if they existed instead of checking for order as well. The other reason I did this is that it doesn't seem like the order in which units are translated is guaranteed by srcML's parser-- I might be wrong, I'd have to ask. But when I ran it myself, the order came out different several times.